LPS-74807 Rename package
