+++ modules/apps/foundation/petra/petra-memory/src/main/java/com/liferay/petra/memory/DeleteFileFinalizeAction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.memory;
+package com.liferay.petra.memory;
 
 import java.io.File;
