+++ modules/apps/foundation/petra/petra-memory/src/main/java/com/liferay/petra/memory/EqualityPhantomReference.java
@@ -12,25 +12,17 @@
  * details.
  */
 
-package com.liferay.portal.kernel.memory;
+package com.liferay.petra.memory;
 
+import java.lang.ref.PhantomReference;
 import java.lang.ref.ReferenceQueue;
-import java.lang.ref.WeakReference;
-
-import java.util.Objects;
 
 /**
  * @author Shuyang Zhou
  */
-public class EqualityWeakReference<T> extends WeakReference<T> {
+public class EqualityPhantomReference<T> extends PhantomReference<T> {
 
-	public EqualityWeakReference(T referent) {
-		super(referent);
-
-		_hashCode = referent.hashCode();
-	}
-
-	public EqualityWeakReference(
+	public EqualityPhantomReference(
 		T referent, ReferenceQueue<? super T> referenceQueue) {
 
 		super(referent, referenceQueue);
@@ -41,17 +33,6 @@ public class EqualityWeakReference<T> extends WeakReference<T> {
 	@Override
 	public boolean equals(Object obj) {
 		if (this == obj) {
-			return true;
-		}
-
-		if (!(obj instanceof EqualityWeakReference<?>)) {
-			return false;
-		}
-
-		EqualityWeakReference<?> equalityWeakReference =
-			(EqualityWeakReference<?>)obj;
-
-		if (Objects.equals(get(), equalityWeakReference.get())) {
 			return true;
 		}
