+++ modules/apps/foundation/petra/petra-memory/src/main/java/com/liferay/petra/memory/EqualityWeakReference.java
@@ -12,17 +12,25 @@
  * details.
  */
 
-package com.liferay.portal.kernel.memory;
+package com.liferay.petra.memory;
 
-import java.lang.ref.PhantomReference;
 import java.lang.ref.ReferenceQueue;
+import java.lang.ref.WeakReference;
+
+import java.util.Objects;
 
 /**
  * @author Shuyang Zhou
  */
-public class EqualityPhantomReference<T> extends PhantomReference<T> {
+public class EqualityWeakReference<T> extends WeakReference<T> {
 
-	public EqualityPhantomReference(
+	public EqualityWeakReference(T referent) {
+		super(referent);
+
+		_hashCode = referent.hashCode();
+	}
+
+	public EqualityWeakReference(
 		T referent, ReferenceQueue<? super T> referenceQueue) {
 
 		super(referent, referenceQueue);
@@ -33,6 +41,17 @@ public class EqualityPhantomReference<T> extends PhantomReference<T> {
 	@Override
 	public boolean equals(Object obj) {
 		if (this == obj) {
+			return true;
+		}
+
+		if (!(obj instanceof EqualityWeakReference<?>)) {
+			return false;
+		}
+
+		EqualityWeakReference<?> equalityWeakReference =
+			(EqualityWeakReference<?>)obj;
+
+		if (Objects.equals(get(), equalityWeakReference.get())) {
 			return true;
 		}
