+++ modules/apps/foundation/petra/petra-memory/src/main/java/com/liferay/petra/memory/SoftReferencePool.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.memory;
+package com.liferay.petra.memory;
 
 import java.lang.ref.ReferenceQueue;
 import java.lang.ref.SoftReference;
