+++ modules/apps/foundation/petra/petra-memory/src/main/resources/com/liferay/petra/memory/packageinfo
